requirements.txt
langchain-core==0.3.60
langchain-openai==0.3.17
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42
python-dotenv
langchain_google_genai
streamlit